Spring Security  - Room Groups, Room Features pages (including Add and Edit pages) changed to use the new security model  - If a user can add both global and departmental room group / feature -- it is now possible to change between the two on the Add page  - Edit page now only contains rooms from the department selected on the Rooms Features / Groups page    - For instance, when looking at final examination rooms, only final examination rooms are shown in the list (and can be added to / removed from the group / feature)

+++ trunk/JavaSource/org/unitime/timetable/model/Department.java
@@ -118,7 +118,7 @@ public class Department extends BaseDepartment implements Comparable, Qualifiabl
 	public static Department findByDeptCode(String deptCode, Long sessionId, org.hibernate.Session hibSession) {
 		return (Department)hibSession.
 			createQuery("select distinct d from Department as d where d.deptCode=:deptCode and d.session.uniqueId=:sessionId").
-			setLong("sessionId", sessionId.longValue()).
+			setLong("sessionId", sessionId).
 			setString("deptCode", deptCode).
 			setCacheable(true).
 			setFlushMode(FlushMode.MANUAL).
